LayoutTests/imported/w3c/web-platform-tests/webrtc/legacy/munge-dont-expected.txt


PASS RTCSessionDescription.type is read-only
PASS RTCSessionDescription.sdp is read-only
PASS RTCIceCandidate.candidate is read-only
FAIL Rejects SDP munging between createOffer and setLocalDescription assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Rejects SDP munging between createAnswer and setLocalDescription assert_unreached: Should have rejected: undefined Reached unreachable code
